# Add prefixed properties to the style object before the unprefixed property