--- a/modules/utils/prefixProperty.js
+++ b/modules/utils/prefixProperty.js
@@ -7,9 +7,20 @@ export default function prefixProperty(
   style: Object
 ): void {
   if (prefixProperties.hasOwnProperty(property)) {
+    const newStyle = {}
     const requiredPrefixes = prefixProperties[property]
-    for (let i = 0, len = requiredPrefixes.length; i < len; ++i) {
-      style[requiredPrefixes[i] + capitalizeString(property)] = style[property]
+    const capitalizedProperty = capitalizeString(property)
+    const keys = Object.keys(style)
+    for (let i = 0; i < keys.length; i++) {
+      const styleProperty = keys[i]
+      if (styleProperty === property) {
+        for (let j = 0; j < requiredPrefixes.length; j++) {
+          newStyle[requiredPrefixes[j] + capitalizedProperty] = style[property]
+        }
+      }
+      newStyle[styleProperty] = style[styleProperty]
     }
+    return newStyle
   }
+  return style
 }