@rofrischmann 2017-12-07
Oh man, you did a great job here! I removed it once because it dropped performance a lot, but now it's totally worth adding that!
Thank you ❤️