@ryan-roemer 2017-12-19
Thanks @rofrischmann !!!

Any chance we can get a new release to npm?